+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] <!-- release-date -->
 
+### Fixed
+
+- Keep rendering while waiting for child processes to shut down.
+
 ## [0.11.0] - 2023-10-21
 
 ### Added